requirements.txt
asgiref==3.6.0
Django==4.0
django-extra-fields==3.0.2
djangorestframework==3.12.4
Pillow==9.4.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
